requirements.txt
langchain
azure-search-documents
azure-identity
python-dotenv
pypdf
openai
langchain-community
uvicorn
fastapi
langchain-openai